+++ frontend/components/commons/persistent-storage/usePersistentStorageViewModel.ts
@@ -1,0 +1,12 @@
+import { useRunningEnvironment } from "~/v1/infrastructure/services/useRunningEnvironment";
+import { useRole } from "@/v1/infrastructure/services";
+
+export const usePersistentStorageViewModel = () => {
+  const { hasPersistentStorageWarning } = useRunningEnvironment();
+  const { isAdminOrOwnerRole } = useRole();
+
+  return {
+    hasPersistentStorageWarning,
+    isAdminOrOwnerRole,
+  };
+};